Rename of file with testing helpers
